chore: Move to theme config
